requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-jose==3.3.0
pydantic==1.10.11
requests==2.31.0
langchain==0.0.500   
openai==0.28.0      
sqlalchemy==2.0.20
databases==0.6.1
python-dotenv==1.0.0
